Emacs erlang-mode: fix syntax highlighting of $ in two cases
A string whose last character is a dollar sign used to make the syntax
highlighter believe that the string never ends, breaking highlighting
of following code:

-vsn("$Revision: 42 $").

And the double quote as a character constant with a (superfluous)
backslash used to make the syntax highlighter believe that a new
string started:

foo() ->
    $\".

This change fixes both problems by adding two regexps to
font-lock-syntactic-keywords in erlang-font-lock-init.

One case that is still broken is when a multi-line string ends with a
dollar sign:

bar() ->
    "This multi-line string
ends with a $".

baz() ->
    this_gets_incorrectly_highlighted.