Add xref support in erlang.el
In GNU Emacs 25 xref will be introduced.  It is a framework for cross
referencing commands, in particular commands for finding definitions.
It does not replace etags.  It rather resides on top of it and
provides user-friendly commands.  The idea is that the user commands
should be the same regardless of what back-end does the actual finding
of definitions.

Add to the xref commands awareness of the module:tag syntax in a
similar way that is already done for the old etags commands.

Xref completion support is not included in this commit.

Remove all compilation warnings for GNU Emacs 24.5 and current 25
(4ffec91).

Remove XEmacs incompatibility in
erlang-font-lock-exported-function-name-face.

Add file erlang-test.el with a single unit test.

Add TAGS to .gitignore.